requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
nose==1.3.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
Werkzeug==0.14.1
